Partition name check for prepartitioned devices
Just skipping the check for a partition named 'ceph' leads to an error
if a devices is given as a precreated partition (eg. to have more than
one OSD on a device). This fixes this case by also checking these
devices for a partiton named 'ceph'.

With the existing logic ceph prepare was run on these devices
unconditionally which lead to an error on subsequent runs after the
device is prepared.